Integrate rstool into Main.
rstool is an aggregation of the existing tools for working with RecordStores.  You can make and merge RecordStores, dump the contents of a RecordStore to individual files, translate between different types of RecordStores, and unhash MD5 digested RecordStore keys given a translation RecordStore.

Also integrates bug fixes to AutoArray and an overloaded mergeRecordStores() function.

[git-p4: depot-paths = "//depot/projects/EVALUATION/Main/tools/src/rstool/": change = 2241]