Add option to "force" insertion into a RecordStore during a key collision scenario.
Pass "-f" to rstool to perform a replace() when the key being inserted already exists in the RecordStore.  This will replace in both the data RecordStore and the hash translation RecordStore.  Without "-f", rstool will cowardly refuse to do the replacement and exit, even if there are more records in the queue.

[git-p4: depot-paths = "//depot/projects/EVALUATION/Main/tools/src/rstool/": change = 4123]